Fix bug where countdown shows incorrect days
If the number of days between the countdown date and the current date
was greater than the number of days in the current month, then the
number of days shown in the countdown timer was incorrect. To fix this,
instead of using the Date.getDate() method, calculate the number of days
explicitly and round the number down.

Fixes #25